requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.8
psycopg2-binary==2.9.9
itsdangerous==2.1.2
Werkzeug==3.0.1
Jinja2==3.1.3
alembic==1.13.1
gunicorn==20.1.0
dependency-injector==4.41.0
python-dotenv==0.21.0
marshmallow==3.5.0
celery==5.3.6
redis==5.0.3